fix: Move chat package
