--- src/main/java/com/cherryhouse/server/chat/ChatController.java
@@ -1,26 +1,0 @@
-package com.cherryhouse.server.chat;
-
-import com.cherryhouse.server.chat.dto.ChatDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.SendTo;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@RequiredArgsConstructor
-public class ChatController {
-
-    private final ChatService chatService;
-
-    @MessageMapping("/chat.{chatRoomId}")
-    @SendTo("/queue/{chatRoomId}")
-    public ChatDto chat(ChatDto chatDto, @DestinationVariable Long chatRoomId) {
-        Chat chat = chatService.create(chatDto, chatRoomId);
-        return ChatDto.builder()
-                .sender(chat.getSender().getId())
-                .message(chat.getContent())
-                .type(chat.getType())
-                .build();
-    }
-}
