+++ src/main/java/com/cherryhouse/server/chatroom/ChatRoomController.java
@@ -2,10 +2,16 @@ package com.cherryhouse.server.chatroom;
 
 import com.cherryhouse.server._core.security.UserPrincipal;
 import com.cherryhouse.server._core.util.ApiResponse;
+import com.cherryhouse.server.chatroom.chat.Chat;
+import com.cherryhouse.server.chatroom.chat.ChatService;
+import com.cherryhouse.server.chatroom.dto.ChatDto;
 import com.cherryhouse.server.chatroom.dto.ChatRoomResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.handler.annotation.DestinationVariable;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,6 +21,7 @@ import org.springframework.web.bind.annotation.*;
 public class ChatRoomController {
 
     private final ChatRoomService chatRoomService;
+    private final ChatService chatService;
 
     @GetMapping
     public ResponseEntity<?> getChatRooms(@AuthenticationPrincipal UserPrincipal userPrincipal, Pageable pageable){
@@ -34,5 +41,16 @@ public class ChatRoomController {
                                       @PathVariable("chatRoomId") Long chatRoomId, Pageable pageable){
         ChatRoomResponse.ChatsDto response = chatRoomService.getChats(pageable, chatRoomId, userPrincipal.getEmail());
         return ResponseEntity.ok().body(ApiResponse.success(response));
+    }
+
+    @MessageMapping("/chat.{chatRoomId}")
+    @SendTo("/queue/{chatRoomId}")
+    public ChatDto chat(ChatDto chatDto, @DestinationVariable Long chatRoomId) {
+        Chat chat = chatService.create(chatDto, chatRoomId);
+        return ChatDto.builder()
+                .sender(chat.getSender().getId())
+                .message(chat.getContent())
+                .type(chat.getType())
+                .build();
     }
 }
