+++ src/main/java/com/cherryhouse/server/chatroom/ChatRoomService.java
@@ -3,8 +3,8 @@ package com.cherryhouse.server.chatroom;
 import com.cherryhouse.server._core.exception.ApiException;
 import com.cherryhouse.server._core.exception.ExceptionCode;
 import com.cherryhouse.server._core.util.PageData;
-import com.cherryhouse.server.chat.Chat;
-import com.cherryhouse.server.chat.ChatRepository;
+import com.cherryhouse.server.chatroom.chat.Chat;
+import com.cherryhouse.server.chatroom.chat.ChatRepository;
 import com.cherryhouse.server.chatroom.dto.ChatRoomResponse;
 import com.cherryhouse.server.post.Post;
 import com.cherryhouse.server.post.PostService;
